Объединяет actions, api-actions.
