+++ src/components/offers-sorting/offers-sorting.jsx
@@ -2,7 +2,7 @@ import React, {useState, useEffect, createRef} from 'react';
 import PropTypes from 'prop-types';
 
 import {connect} from 'react-redux';
-import {ActionCreator} from '../../store/main/action';
+import {ActionCreator} from '../../store/action';
 import {getActiveSorting} from '../../store/main/selectors';
 
 import {SortingTypes} from '../../const';
