+++ src/components/sign-in-form/sign-in-form.jsx
@@ -2,7 +2,7 @@ import React, {useRef} from 'react';
 import PropTypes from 'prop-types';
 
 import {connect} from 'react-redux';
-import {login} from '../../store/user/api-actions';
+import {login} from '../../store/api-actions';
 
 const SignInForm = (props) => {
   const {onSubmit} = props;
