+++ src/store/favorites/reducer.js
@@ -1,4 +1,4 @@
-import {ActionType} from './action';
+import {ActionType} from '../action';
 import {adaptOffersData} from '../../services/adapter';
 
 const initialState = {
