--- src/store/user/action.js
@@ -1,15 +1,0 @@
-export const ActionType = {
-  REQUIRED_AUTHORIZATION: `user/requiredAuthorization`,
-  SET_AUTH_INFO: `user/login`
-};
-
-export const ActionCreator = {
-  requireAuthorization: (status) => ({
-    type: ActionType.REQUIRED_AUTHORIZATION,
-    payload: status,
-  }),
-  setAuthInfo: (info) => ({
-    type: ActionType.SET_AUTH_INFO,
-    payload: info
-  })
-};
