--- src/store/user/api-actions.js
@@ -1,25 +1,0 @@
-import {ActionCreator} from './action';
-import {ActionCreator as middlewareActionCreator} from '../middlewares/action';
-import {AppRoutes, APIRoutes, AuthorizationStatus} from '../../const';
-
-export const checkAuth = () => (dispatch, _getState, api) => (
-  api.get(APIRoutes.LOGIN)
-    .then(({data}) => dispatch(ActionCreator.setAuthInfo(data)))
-    .then(() => dispatch(ActionCreator.requireAuthorization(AuthorizationStatus.AUTH)))
-    .catch(() => {})
-);
-
-export const login = ({email, password}) => (dispatch, _getState, api) => (
-  api.post(APIRoutes.LOGIN, {email, password})
-    .then(({data}) => dispatch(ActionCreator.setAuthInfo(data)))
-    .then(() => dispatch(ActionCreator.requireAuthorization(AuthorizationStatus.AUTH)))
-    .then(() => dispatch(middlewareActionCreator.redirectToRoute(AppRoutes.MAIN)))
-    .catch(() => {})
-);
-
-export const logout = () => (dispatch, _getState, api) => (
-  api.get(APIRoutes.LOGOUT)
-    .then(() => dispatch(ActionCreator.requireAuthorization(AuthorizationStatus.NO_AUTH)))
-    .then(() => dispatch(ActionCreator.setAuthInfo({})))
-    .catch(() => {})
-);
