+++ src/store/user/reducer.js
@@ -1,4 +1,4 @@
-import {ActionType} from './action';
+import {ActionType} from '../action';
 import {AuthorizationStatus} from '../../const';
 
 const initialState = {
